pyproject.toml
[tool.poetry]
name = "market-research"
version = "0.1.0"
description = ""
authors = ["Bruno Tatsuya <brunomasunaga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
inject = "^5.2.1"
pydantic = "^2.8.2"
pyfunctional = "^1.5.0"
sqlmodel = "^0.0.19"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
typer = "^0.12.3"
pytest = "^8.3.2"
pydantic-settings = "^2.3.4"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
